Corrections for open addresses Signed-off-by:Bart Hanssens <bart.hanssens@bosa.fgov.be>

+++ java/dao/src/main/java/be/bosa/dt/best/dao/Geopoint.java
@@ -48,7 +48,7 @@ public class Geopoint {
 	 * Set X coordinate
 	 * @param x 
 	 */
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
